Split Map.jsx into ControlPanel.jsx sidebar
