+++ src/Map.jsx
@@ -6,8 +6,6 @@ import {
   createRefetchContainer,
 } from 'react-relay';
 import propTypes from 'prop-types';
-import { DateTimePicker } from 'react-widgets';
-import Toggle from 'react-toggle';
 import * as turf from '@turf/turf';
 import { MAP_STYLE, MAPBOX_ACCESS_TOKEN } from './config.json';
 import {
@@ -16,43 +14,7 @@ import {
   getVehicleMarkersLayer,
   getSubRoutesLayer,
 } from './Route';
-import muniRoutesGeoJson from './res/muniRoutes.geo.json';
-import Checkbox from './Checkbox';
-
-const notAlpha = /[^a-zA-Z]/g;
-
-/*
-Sort by putting letters before numbers and treat number strings as integers.
-Calling Array.prototype.sort() without a compare function sorts elements as
-strings by Unicode code point order, e.g. [2, 12, 13, 9, 1, 27].sort() returns
-[1, 12, 13, 2, 27, 9]
-*/
-function sortAlphaNumeric(a, b) {
-  const aRoute = a.properties.name;
-  const bRoute = b.properties.name;
-  const aInteger = parseInt(aRoute, 10);
-  const bInteger = parseInt(bRoute, 10);
-  const aAlpha = aRoute.replace(notAlpha, '');
-  const bAlpha = bRoute.replace(notAlpha, '');
-
-  // special case for K/T and K-OWL
-  if (aRoute === 'K/T' && bRoute === 'K-OWL') return -1;
-  if (aRoute === 'K-OWL' && bRoute === 'K/T') return 1;
-
-  if (Number.isNaN(aInteger) && Number.isNaN(bInteger)) {
-    return aAlpha < bAlpha ? -1 : 1;
-  } else if (Number.isNaN(aInteger)) {
-    return -1;
-  } else if (Number.isNaN(bInteger)) {
-    return 1;
-  } else if (aInteger === bInteger) {
-    return aAlpha < bAlpha ? -1 : 1;
-  }
-  return aInteger - bInteger;
-}
-
-// make a copy of routes and sort
-const sortedRoutes = muniRoutesGeoJson.features.slice(0).sort(sortAlphaNumeric);
+import ControlPanel from './ControlPanel';
 
 /*
 * Stop class used to handle info about selected stops
@@ -84,6 +46,7 @@ class Stop {
 class Map extends Component {
   constructor() {
     super();
+    this.filterRoutes = this.filterRoutes.bind(this);
     this.state = {
       // Viewport settings that is shared between mapbox and deck.gl
       viewport: {
@@ -99,7 +62,6 @@ class Map extends Component {
         coordinates: { lon: 0, lat: 0 },
         info: { vid: '', heading: 0 },
       },
-      currentStateTime: new Date(Date.now()),
       showStops: true,
       selectedStops: [],
       subroute: null,
@@ -116,23 +78,6 @@ class Map extends Component {
     window.removeEventListener('resize', this.updateDimensions.bind(this));
   }
 
-  setNewStateTime(newStateTime) {
-    this.setState({ currentStateTime: newStateTime });
-    this.props.relay.refetch(
-      {
-        startTime: Number(newStateTime) - 15000,
-        endTime: Number(newStateTime),
-        agency: 'muni',
-      },
-      null,
-      (err) => {
-        if (err) {
-          console.warn(err);
-        }
-      },
-      { force: true },
-    );
-  }
   /**
    * given the two selected stop sids, returns a line segment
    * between them
@@ -222,40 +167,6 @@ class Map extends Component {
     this.setState({ geojson: newGeojson });
   }
 
-  renderControlPanel() {
-    return (
-      <div className="control-panel">
-        <div className="routes-header">
-          <h3>Time</h3>
-          <DateTimePicker
-            value={this.state.currentStateTime}
-            onChange={newTime => this.setNewStateTime(newTime)}
-          />
-        </div>
-        <div className="routes-header stops-toggle">
-          <h3>Stops</h3>
-          <Toggle
-            defaultChecked={this.state.showStops}
-            onChange={() => this.setState({ showStops: !this.state.showStops })}
-          />
-        </div>
-        <div className="routes-header">
-          <h3>Routes</h3>
-        </div>
-        <ul className="route-checkboxes">
-          {sortedRoutes.map(route => (
-            <Checkbox
-              route={route}
-              label={route.properties.name}
-              handleCheckboxChange={checkedRoute => this.filterRoutes(checkedRoute)}
-              key={route.id}
-            />
-          ))}
-        </ul>
-      </div>
-    );
-  }
-
   renderMap() {
     const onViewportChange = viewport => this.setState({ viewport });
     const { trynState } = this.props.trynState || {};
@@ -321,7 +232,7 @@ class Map extends Component {
             {this.renderMap()}
           </div>
           <div className="col-sm-3 col-md-2 hidden-xs-down bg-faded sidebar">
-            {this.renderControlPanel()}
+            <ControlPanel filter={this.filterRoutes} />
           </div>
         </div>
       </div>
@@ -334,7 +245,6 @@ Map.propTypes = {
     propTypes.string,
     propTypes.arrayOf(propTypes.object),
   ).isRequired,
-  relay: propTypes.element.isRequired,
 };
 
 export default createRefetchContainer(
